viewpaint_print_addr(): Change off parameter to UINT32
That took me a while to notice that the view1mb addresses were cut off after 0xffff... -.-

+++ win9x/debuguty/viewpaint.cpp
@@ -44,7 +44,7 @@ COLORREF viewpaint_line_set_colors(NP2VIEW_T *view, RECT *rc, HDC hdc, DWORD y, 
 	return bkcol;
 }
 
-void viewpaint_print_addr(HDC hdc, int x, int y, UINT16 seg, UINT16 off, BOOL segmented)	{
+void viewpaint_print_addr(HDC hdc, int x, int y, UINT16 seg, UINT32 off, BOOL segmented)	{
 
 	TCHAR	str[16];
